x86/testing/cgroups_tests: Test cases for existing cgroup infrastructure
These are the initial test cases as suggested by Fenghua, to test the
existing cgroup infrastructure. These have a couple of nasty bugs and
some optimizations that can be done, I will work on that. Posting them
now so that they can be used. Please go through README file which
explains the way to run these tests and to make sure that you dont
trigger bugs.

Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>